Auto-mark as internal and compatibility with all graphs
Neat functions are now auto-marked as internal only on startup. This allows UFUNCTION developers to only mark them as Neat, which means functions can be used properly even if the plugin is disabled. The old setup would force the developer to make a decision whether to expose the regular function or not, which would either cause duplicate functions to be visible, or their function to not be available if an end user didn't have the plugin enabled.

The call function node now works in all contexts, so if you're using it in an event graph it looks neat but if you use it in a function graph it will not do anything, it will have delegate pins on the input side like a regular function would.