Autowire now connects to delegate's Then pin
Basically, when creating a connection between a data pin on a Neat function node and some other node, Unreal tries to connect the Then pin of the Neat function with the Exec pin of the other node. This is generally incorrect for any delegates, since the values from the data pins are not necessarily valid in that context.

Instead, we redo the autowire, trying to connect to the corresponding delegate's Then pin instead.